prepare tab container for new theme
