+++ src/app-pages/project/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import SecondaryNavBar from '../../app-components/navigation/secondaryNavBar';
+
+const Project = () => (
+  <div>
+    <SecondaryNavBar
+      headingNode={<h4>Huntington District</h4>}
+    />
+    Project Page!
+  </div>
+);
+
+export default Project;
